Repositories View: wrong buttons when showing Remote Configuration Properties
The Properties page exposed by the Repositories View is reused between
showing the repository configuration and the remote configuration(s) (the latter
is shown when a "Remote" node or one of its children is selected in the view).
Currently, the toolbar is polluted with unwanted buttons when switching from
a repository config to a remote config.

Change-Id: Ic27288d8b80cb634ac4cf2629c9bcf1947a7babc
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>