Update index when Project is moved
Do the same as was done before when a folder is moved:
Move the index content to the new location. The index then has no
obsolete entries. All moved resources known to the index are marked
as added afterwards and can easily be committed.

There is still an issue that the deletion of the old resources cannot be
committed from egit (bug 302349). This applies also to the deletion of
projects and is not fixed by this commit.

Bug: 297980
Change-Id: I5b912d52f6c9e481ab6e1e2a9038d6d4c8d7d72e
Signed-off-by: Stefan Lay <stefan.lay@sap.com>