Support Repository Creation
Repository creation is supported either by calling a "New Wizard"
(CTRL-N), then Git -> Git Repository or by using a new action in the
Git Repositories View.
The dialog asks for a directory in the local file system and whether
the repository is to be bare. The repository is added to the Git	
Repositories View automatically.
In the course of this, the toolbar and view menu of the Repositories View
was also cleaned up.

Bug: 315015

Change-Id: I65037930e918b3252a27cb27f4ca1c47d9c9cf10
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>