Update patch set 5
Patch Set 5:

When a synchronization request is sent, another job for processing subscriber change events are scheduled. This job, while updating the same label as the job that the synchronization request is performed on (which makes it seem like there is actually only one job running), does not appear to share the same progress monitor.

Hence, the synchronization job returns early while the subscriber event processing job runs to completion.

Patch-set: 5