Git Repositories View: remote configuration handling
Remote configuration handling in the Git Repositories view must
behave like the standard: it is possible to push to a remote repository
if only a "url" is present in the config, but no "pushurl". If both
url and pushurl or pushurls exist, they are all used.
Accordingly, the "configure push" UI must show the (fetch-) URL
if it is present and make addition of "pushurl"s optional in this case.
Also the Repo View content and label providers had to be adjusted in 
order to show the "push" node if only a "url" and a "push" but no 
"pushurl" is defined.

Change-Id: Ia1e44002c2949a243326798c182a26c4caad56d2
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>