Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

When renaming a remote branch, the result is a local branch (when trying to rename refs/remaotes/origin/master, the new branch will be refs/heads/<new name>)...
Also, there is a NullPointerException when renaming a remote branch. This is probably a bug somewhere else, as there is also a NPE when renaming a remote branch from the Team->Branch... menu.

Patch-set: 2