Git Repositories View: Hierarchical branches representation toggle
Similar to the package explorer which supports flat vs. hierarchical
package representation, this allows to toggle between flat vs. hierarchical
branch representation. If the toggle is on, then the "/" used in the 
branch names is interpreted as hierarchy separator and the branches are
presented correspondingly.
Other than the suggestion in bug 315567, this does not eliminate the
"Remote Branches" node altogether in "hierarchical" mode, since 
there might be remote branches without "/" separators 
which would result in a confusing representation.

Bug: 315567
Bug: 323950

Change-Id: I085dc24a2cd517a76b9d210c9f74d2832dfe9ff4
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>