Update patch set 1
Patch Set 1:

Not sure if this is really needed: the currently checked out branch is visualized now on the Repository node in brackets similar to the visualization in the Git Label decorator...
I just saw that the heuristics of "guessing" the remote branch or tag in case of detached HEAD is missing from this visualization, but it should be easy to add it similar to the label decoration. 
Hence, wouldn't we make this view unneccessarily complex by using upward inheritance for the "checked out" decoration?

Patch-set: 1