Supportability: Provide a Trace Configuration Preference Page
Since we have the infrastructure in place, we can now switch debug
tracing on and off in a running IDE. 
This can be very helpful in situations where we can't debug.
Unfortunately, there is no generic UI to do this,
so this adds a preference page to show the existing trace locations for
the EGit UI and EGit Core plug-ins and allows to activate them.
This page is generic enough to be easily extended to include the 
locations of other plug-ins as long as they use an .options file to declare
the locations (and don't do anything else in this .options file).

Change-Id: I4681acf85de8d05be6290d88acde5d99453ce89d
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>