History View: always show selected resource
The behavior of the history view is quite confusing for most users:
instead of showing the current input, a combination of input and
the currently checked button is displayed above the list. Not only
does the user have a hard time to understand what they are seeing
and why they have certain context menus available at a given time,
but also when the "Repository" button is selected, the repository name
is always shown as input, no matter where you click. Subsequently,
the navigation history will contain the same String several times, and
you don't see what you have in fact visited.
This change replaces does replace the current String with something
like "File: MyProject/folder/File.txt [myRepository]", or
"Project: MyProject [myRepository]" etc.
The getDescription() method was also changed to return something
like the above appended with a description of the current "filter" setting,
for example "All changes in Repository of this Resource" or "All changes
of this Resource", but unfortunately, this information is never rendered
(the javadoc says it might be rendered as tooltip or such)... I think I'll
open a bug for this one....
This also fixes some missing mnemonics in the view menu.

Bug: 320998
Change-Id: I9409728d133a70875f9ee16ff93cce33a94361fb
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>