Adds support for comparing two remote branches
The main reason of not displaying synchronization data for two remote
branches was that GitResourceVariantTreeComparator was using standard
SyncInfo class. With uses local file to determinate change kind.

The new GitSyncInfo is looking for a first diff for given file in given
range of commits. Based on this it determinate change kind and
direction.

Now we can display differences between two remote branches also in
ChangeSet and in normal synchronization mode.

Bug: 324044
Change-Id: Ia0cf491657a6c9a01d68992457a25d15c5718428
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>