Adds support for comparing two remote branches
The main reason of not displaying synchronization data for two remote
branches was that GitResourceVariantTreeComparator was using standard
SyncInfo class which uses local file to determine change kind.

The new GitSyncInfo is looking for a first diff for given file in given
range of commits. Based on this it determines change kind and
direction.

Git specific SyncInfo is activated only when the user doesn't select
"include locally changed files" in synchronization dialog. When this
option is selected standard implementation of SyncInfo will be used.

Additionally "Workspace" model is only enabled when "include locally
changed files" option is selected, in other case this model will be
excluded.

Bug: 324044
Change-Id: Ia0cf491657a6c9a01d68992457a25d15c5718428
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>