History View: fix view menu
Eclipse UI Guidelines mandate that all toolbar actions should also appear
in either the toolbar menu or a context menu.
This adds the toolbar items to the view menu where possible and also
adds specific menu label texts with mnemonics instead of the lengthy
tooltips currently re-used in the menu.

Change-Id: I068a6dcc599fe7d13810d1e2519cb8fa578cefa0
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>