Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The text fields for merge result and new head look like they were editable but they are not (which makes sense). It should be visible that these fields are read-only.

I think the merge inputs should be shown above the result since the operation starts from the inputs to result in the merge result.

Would be also interesting to see which branch or tag has been merged into which branch (if a branch or tag has been merged).

It's not really obvious that the entries in the merge input table are representing commits since the column labels are rather generic.

Patch-set: 2