Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

When calling create branch from the repository view, the validator is asked to validate refs/heads/refs/heads/<myInput> instead of refs/heads/<myInput>.
We also have a little usbility issue if this is not used directly in an InputDialog. For example if you open the tag dialog (either from team or repo view context menu), the tag name will be shown as "invalid ref name", even though the user hasn't entered anything yet. Looks like we need to distinguish the two cases in some way or another: if the validator is used outside an InputDialog, the validator should accept empty names as ok. Might be a new method in ValidationUtils or something.

Patch-set: 2