Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I have to second Chris here, I also don't think this should be (at least in this shape) part of the regular install. What I could imagine is that the preference page is not visible by default and we can fire it up with a command. Using a command for this has the advantage that everyone can easily execute it via Ctrl+3 but it doesn't clutter the UI in any way. Opinions?

Patch-set: 2