Temporary revert refactoring in GitModelObject.createTreeWAlk()
My refactoring in commit 4baa618 breaks ChangeSet displaying because
FileTreeIterator is used to determinate ignored files. We should
implement a variant of WorkingTreeIterator that will work on given tree
instead of actual working file. I'm afraid that we can't made such
implementations before 0.9 release therefore we should revert this
change and work on it after release.

Bug: 324705
Change-Id: I6623c5d48e797bb125d7697ce778eaf628e0a2b8
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>