Update patch set 1
Patch Set 1: IP review completed

As Jens mentioned, we have an issue when going through the Git Import wizard. We have to see whether we should close the wizard and pop it back up after the clone is done. If we have a checkbox saying whether or not the clone should be done in the background, we can decide how to act. Not all clones will take forever and warrant the running in the background behavior.

Patch-set: 1