Fix NullPointer in HistoryView
With 7f6821cfd3b577bbb82f12e54f36cd8e6d1904da a NullPointer was
introduced.  This is fixed with this commit.

Now currentWalk is not set to null per default and it is only 
released if necessary. Because cancelRefreshJob() may set currentWalk 
to null it has to be newly created in case it is null.

Bug: 324943
Bug: 325081
Change-Id: I0d6c032e35576ded39d4bced1ff5dfde613cc932
Signed-off-by: Stefan Lay <stefan.lay@sap.com>