Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(1 inline comment)

Now the Git Change Set Model works fine for comparing two arbitrary branches.

But on the other hand, the workspace model is now broken when the HEAD is synchronized with another branch. In the workspace model the left side is the workspace version of the file. Now uncommitted changes are not visible anymore in the synchronize view.

Furthermore the workspace model is still broken in case of comparing two arbitrary branches (see bug 322489). Could it be switched off in that case?

Patch-set: 2