Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

I would like to switch off the workspace model in case of comparing two arbitrary branches because it compares the workspace with the remote state. I suppose that here the team API must be extended.

But the main issue with this change is that the workspace model is now broken when the HEAD is synchronized with another branch. Now I do not see the local changes anymore, even if I check the "Include local..." checkbox.

Patch-set: 3