Update patch set 1
Patch Set 1:

Hi Robin,

Some comments on this:

1. RCS detects changes in the index. Decorator reacts to these changes (e.g. staged, conflict).

2. RCS also triggers RIRefresh to refresh Eclipse resources.
You mention this happens faster than when it's done by the standard Eclipse mechanism. Is this the right approach?
Shoudn't all external changes to files be handled in the same way (reaction time) regardless wheter the files relate to resources in a Git repository or not?
With the current implementation you have the following effect if Workspace auto refresh is enabled:

  1. change a file in a project shared with Git using Notepad => immediate refresh of the Eclipse resource
  2. change a file in a project NOT shared with Git using Notepad => refresh of the Eclipse resource after some delay.

Patch-set: 1