Update patch set 1
Patch Set 1:

The first version was independent of the workspace refresh job, but upon request I changed it. If you have automatic refresh you probably want it always be refreshed, but scanning all the time affects performance too much, so Eclipse scans with an interval.

Ideally we should have a separate option for this and it should also interact with or replace the standard job so we don?t have two of them).

Patch-set: 1