Update patch set 1
Patch Set 1:

Not so fast! I admin this code is so ugly, it should probably be deleted for that reason, but there is difference
from what the general refresh does and that is that this code reacts to changes in the index and triggers a refresh earlier than the standard workspace refresh, which the reason Egit reacts quickly when you do something to the repo outside of Eclipse. Without this you have to refresh manually all the time which you (read: I) will forget all the time. It is the RCS job that triggers then index change refresh.

A possible cleanup action would be to kick the standard workspace Job, if possible when the index changes. When I wrote this code it was not accessible.

Patch-set: 1