Update patch set 3
Patch Set 3:

Scanning the full workspace is heavy for big repos.If you want full refresh for any resource change use the workspace refresh. This refresh can works even without hitting the physical disk since it reads so little that the disk cache is likely to be used, which is good for the battery on my laptop.

I've seen no problem with the decorators, but you and others have reported such pronblems and I think this patch does not change this.

Patch-set: 3