Update patch set 2
Patch Set 2:

Hi Ketan,
yes, that's the problem. I had the issue that an item was not loaded because it was not in visible.
I could increase the size of the view to make the item visible, but I think that's no stable solution. In my workaround I forced loading the TableItem.

Patch-set: 2