Add commit hash to history view
When you compare two revisions from the history it showed the
commit hash next the file name but from the history view table it was 
not clear which revision is which. To find out which revision is which
you needed to select a single revision from the history view and
then check the detail for that commit.

Bug: 327291
Change-Id: I073ed81ca022fc22779accff75fc56d7ffb98dfc
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>