CreateTagDialog: improve UI Responsiveness
This improves UI responsiveness of the CreateTagDialog by filling the
"Existing Tags" list asynchronously and by filling the "Commit" combo
box lazily when opening the "Advanced" section.
While the tag list is filled, a "Loading..." node is shown instead and the
list is inactive. Still it is possible to use the dialog while the list is filled.
In addition, the content provider for the tag list was changed to a simple
ArrayContentProvider which saves some extra overhead.

Bug: 324108
Change-Id: Ic1925916e0131a5aee96d92fe56053f0f5ec4f2f
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>