Update patch set 1
Patch Set 1: (3 inline comments)

I haven't actually tested the code but it seems okay. I would have considered a) using DeferredTreeContentManager but that might be a little bit much or b) having the tags retrieved before the population of the dialog but I think the "Loading..." message should be sufficient.

If you want to refine the behaviour even further, you could add support for progress monitoring and reporting...though I realize you've set it up to be a system job at the moment (though system jobs can be displayed in the 'Progress' view actually). If progress monitoring was implemented then the job could support cancellation properly when the user just closes the dialog before the table actually gets populated. As it is right now, all the tags are retrieved even if the dialog has already been closed.

Patch-set: 1