[historyView] Enable multiple Resource selection
Enable Team->Show in History on a multi-selection of Resources. The
selected resources will be used as filder for the history view.
This does NOT work for the generic Show In->History action; also, if
"Link with current selection" is on, this will always work for the last
clicked resource. The latter limitations are due to a shortcoming of the
history framework.

Change-Id: I33a0acb32af3620aa00ca28693b8c09a7a59d042
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>