Allow to open History View from Repositories View
In that case, the filter icons are disabled (as we don't have a
resource filter to apply).
In the course of this, the ResourceList object was removed from
EGit core and replaces with a package-local HistoryPageInput.
Furthermore, the GitHistoryPageSource was moved to the
correct package and the createPage logic was fixed (the old
logic would always do a setInput on the new page which was
double work as the history framework would to this again).

Bug: 327345
Change-Id: Iee3df7e6fe12f928c8706bd296b38ecaee6326dd
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>