Update patch set 1
Patch Set 1:

The number of selected files is irrelevant here, since the history framework does not work with multi-selection (see comment in GenericHistoryView).
The history view input is always a single resource.
The "Open in Text Editor" behaves exactly like the "Open" in this respect.
You may want to look at 
http://egit.eclipse.org/r/#change,1729
which proposes a workaround for this limitation.

Patch-set: 1