Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; IP review completed

Works fine. I just wonder if we should offer this without also offering an opportunity to search for the ID. From the coding, it doesn't look to difficult to add this.
I think users would become frustrated if they can see the ID but not search for one...

Patch-set: 1