Allow to open History View from Repositories View
In that case, the filter icons are disabled (as we don't have a
resource filter to apply).
In the course of this, the ResourceList object was removed from
EGit core and replaced with HistoryPageInput.
Furthermore, the GitHistoryPageSource was moved to the
correct package and the createPage logic was fixed (the old
logic would always do a setInput on the new page which was
double work as the history framework would to this again).
A better implementation of setInput() results in far fewer calls
to setInpu() (the history framework will only call setInput if the
new object is different from the previous input, but the existing
implementation would return something different upon getInput
than was provided during setInput).
The All Branches and Compare Mode buttons were changed 
to use Preference Actions so that their toggle state survives a 
restart now.
The rest of the changes is adjusting the handlers to deal with
File as input (instead of IResource).

Bug: 327345
Change-Id: Iee3df7e6fe12f928c8706bd296b38ecaee6326dd
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>