[historyView] Refactoring second part
Use the HistoryPageInput as input in the GitHistoryPage. Calculate the
name only when the input changes. Add an adapter so that selection
in the history view automatically changes the input of the history view
if "link with selection" is active. Prepare HistoryPageInput to also allow
to use Files instead of IResources.