[historyView] Refactoring third part
Add a command to allow "Show in History" for multi-selection in the
Repositories View. Adjust all handlers to work with a list of Files as well
as a list of IFiles. Show an error message instead of a wrong history
when the user selects a file that is in the GIT directory.

Change-Id: Idd068ce9e36e1136ebd2b7ec693c7f4a828aaaa9
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>