[historyView] Refactoring fourth part
Replace "Show all Branches" and "Compare mode" actions with
BooleanPrefActions and add the corresponding preferences so that
the toggle state of these buttons survives a restart.
Decouple the different sub-components (file list, diff viewer)
of the history page by using the Preferences directly instead 
of setting the preference values in the components programmatically.