[historyView] Refactoring last part
Make sure the new BooleanPrefActions are disposed properly. Fix the
BooleanPrefAction to avoid executing every action two times instead
of just once. Fix a bug with the "Show Revision Details" action. Fix
an Exception with CompareUtils in case of bare Repositories.
Add a Trace location for the history view and use it.
Display a "Building commit list..." message while waiting for the
list to be created in the background.

Change-Id: I04a39eebbe287aa91e9f186a374cd0f94426a4dc
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>