[historyView] Refactoring last part
Fix the BooleanPrefAction to avoid executing every action two times 
instead of just once. Fix a bug with the "Show Revision Details" action.
Fix an Exception with CompareUtils in case of bare Repositories.
Add a Trace location for the history view and use it.
Display a "Building commit list..." message while waiting for the
list to be created in the background. Fix the issue described in the
original change 1753:

* show history for first repository
* in repositories view select second repository and click "Open in History"
* the History View gets the focus but the history is not switched
to the one for the second repository, instead still the history
of the first repository is displayed.

Bug: 327345
Change-Id: I04a39eebbe287aa91e9f186a374cd0f94426a4dc
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>