Let Repositories View show detached HEADs and special refs
Repositories view was displaying HEAD under "Symbolic References".
But it did never show a detached HEAD. Also the special refs
MERGE_HEAD and FETCH_HEAD are never displayed. This is fixed
by not displaying symbolic references anymore but all the
references which are not shown under Branches and Tags. Maybe
we should also rename the tree node to "References" to tell
that not only symbolic references are shown here.

This bug depends on Change 1772. Don't merge this before we
have merge this (JGit) change

Change-Id: I475aed52ee4eeca530d5671933c4a9e26bf87d6a
Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>