Update patch set 1
Patch Set 1:

I recommend to always have the gerrit Change-Id in the commit message, this helps to avoid trouble when pushing a second patch-set. You may either generate it using the Change-Id checkbox in EGit (hope you commit using EGit :-) or by installing the commit message hook as described in the contributor guide. To update this change manually copy the Change-Id tag from gerrit into the amended commit before pushing it to gerrit. Otherwise you will end up with a second change in review instead of a second patchset for the same change.

Patch-set: 1