Optionally hard-wrap commit message
When the user presses "Commit", the commit message is hard-wrapped to
70 characters if the "Hard-wrap commit message" preference is set.
Existing newlines are not touched, to preserve manual formatting
(e.g. ASCII graphics).

The editor shows the commit message in a configurable font that
defaults to the text font (usually monospaced). The width of the text
field is fixed to 70 characters. This way, what the user sees should
be what the user gets when we later hard-wrap the commit message.

Change-Id: Ic8ae2f99b595f5f994af3cdb08b9363d2d53e685