Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I have prepared a change which also renames the parent to
"Named References" and allows for checking out and creation on
branches on these references.
This is independent from the JGit changes, and I'll upload this to Gerrit as soon as I have access again.

Patch-set: 1