[RFC] Start integration of a merge tool
At the moment, this is intended as Request for Feedback.

This change adds a "Merge Tool" action to the team menu which is
only active if the repository containing the selected resources
is in state "Merging". When selected, a dialog will appear asking
whether the workspace (pre-merged) or HEAD (un-merged) version
of the files should be used as input for the merge editor.

The selected files will be shown in the upper part of the editor
(only the files which have conflicts or have different content in
the workspace compared to the  HEAD version). On double-click, the
actual merge editor will be shown. The user can now edit the left
part and save the changed content to the workspace. Clicking
next/previous the user can navigate to the next/previous conflict/
change (even across file boundaries).

Change-Id: I6d931c24a7122466b566fa61104ea2dbdf72356e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>