AdaptableFileTreeIterator: handling of nested / closed projects
In case of nested projects AdaptableFileTreeIterator finds
more than one container for a given file. The old implementation just
took the first found container. This could cause problems if the
container is not accessible (project closed). The new implementation
does the following:
1. Non accessible containers are excluded
2. The remaining containers checked for the containing projects:
The container contained in the innermost  project is chosen.

Bug: 328489
Change-Id: Iac58c50914e9b9fcafba9f0e0fc4d08a33455ccd
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>