RFC: Start integration of a merge tool
At the moment, this is intended as Request for Feedback: this adds a
"Merge Tool" action to the team menu which is only active if the
Repository of the selected resources is in state "Merging". When
selected, a dialog will appear to ask the user whether they want to
use the workspace (pre-merged) or HEAD (un-merged) version of
the conflicting files as input for the merge editor.
The selected files will be shown in the upper part of the editor (only
such files who have conflicts or whose content differs between the
workspace and HEAD version). On double-click, the actual merge editor
will be shown. The user can now edit the left part and save the changed
content to the workspace. With next/previous they can navigate to the
next/previous conflict/change (even across file boundaries).
The choice on the first dialog is only relevant for conflicting files,
i.e. files that are not auto-merged. In the case of a conflict-free merge,
always the workspace content will be shown on the left side.
The right side will either show the version found at MERGE_HEAD (in
case of conflicts) or HEAD (in case of conflict-free merge).

Change-Id: I6d931c24a7122466b566fa61104ea2dbdf72356e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>