[historyView] Refactor menu contributions
This moves the context menu handling out of the GitHistoryPage
into the respective controls (commit graph table, commit file viewer,
and commit details viewer, respectively).
This reduce the complexity of the GitHistoryPage, and it puts the
responsibilities where they belong.
The new menu contributions discussed in the bug below were also
added to the CommitFile viewer context menu.

Bug: 323072
Change-Id: I6b4579a687894c0ad43d68f4fa4f5c6e3222501f
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>