Support state [removed, untracked] in commit dialog
Currently files which are removed from the index but exist in the
working directory are displayed as "removed" in the commit dialog.
Committing the change removes the affected files from the repository.
Now the files are displayed as "removed" and "untracked"
and committing results in committing a change to the affedted file.

Bug: 327087
Change-Id: I5e789ede05d9bf0487779ef023dbdc90a3197571
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>