Update patch set 1
Patch Set 1:

Robin, we should not clutter the Commit dialog with too many things. As of now, we already have so many options in the commit dialog that commiting is more than a no-brainer. I think this should go to the preference page (same would be for print margin and font). But we may want to provide a link to the pref page in the dialog to quickly access the available options.

Patch-set: 1