Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(3 inline comments)

The editor area for the commit message looks ugly since
- it doesn't align with the file table. Either center it or align it with the file table.
- What's the purpose of the vertical bar at the right end of the commit message pane ?

Whitespace in front of lines in the middle of a paragraph is erased. Could this be preserved to allow indenting inside of paragraphs.

Would be nice to get a warning if there is no blank line after the header, also when the header becomes too long.

Patch-set: 2