[historyView] Added preference for reusing compare editors
It is now configurable whether the same compare editor is used each
time, or a new one is opened each time the user wants to compare
two revisions of a file.

This behavior was changed recently from opening multiple editors to
opening a single one, but personally I like it much better to open
multiple editors.  This makes it easier for me to review all changes
in a commit.

Change-Id: Ib9df76630aa6f59bd5c095f2c5c2fdfc0ddcf283