[historyView] Fix issue with "Folder" filter
When selecting a project in the project explorer and then changing
to "Folder" filter in the history view, the "Building commit list..."
message is shown and not replaced with the commit list. The underlying
reason is that the repository relative path is not calculated
correctly and subsequently the commit list is always empty which in
turn is something the logic that updates the UI is not prepared to
deal with.

Change-Id: Id15c31eb9c552b226a546e1a589078ad583a691c
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>