Fix Git Command Group actions
Make the Git Team Actions implement IWorkbenchWindowActionDelegate so
that the actions in the Git Command Group work again.
This is a prerequisite for fixing bug 329963. The complete fix should
involve replacing the icons with the new ones (and removing the old
icons), adding the "definitionId" so that the keyboard shortcuts work
(which probably should allow us to close a few other bugs) and
thinking about additional actions to add here (e.g. pull, rebase...).

Change-Id: I23895129bfbcd86c70ebfa82fb12839b241c9bf8
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>