Git toolbar and menu using command framework
The old git toolbar and menu were based on actions and didn't use
icons consistent with the team menu. Also some tooltips were missing.
The git actionset can be used to enable the git menu and toolbar.

RepositoryActionHandler was enhanced to register for selection
changes so that command enabling also works for the git toolbar.

Bug: 329963
Change-Id: Ife185129482ea92c8fd5919219c6fb39e1debb8b
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>