Initial support for rebase
This is a first implementation of Rebase. Users can select a branch in
the Repositories View and have a menu item "Rebase". The currently
checked out branch will be rebased onto the selected branch.
If no branch is currently checked out, the menu action will not be
available.

In case of merge conflicts, the rebase will be aborted, i.e. the
original HEAD will be restored and the user will receive an error
pop-up. This is because we found several problems if the repository is
left in state "merging:
1. It is not possible to resolve the conflicts using "add", so the
user can not continue with the rebase (in the command line), no matter
whether they "add" in EGit or in the command line.
2. There are some issues when "aborting" with the command line a
rebase that was started with EGit: HEAD is restored to the
wrong branch (i.e. the target branch)

These issues will be fixed in subsequent changes.

Bug: 328219
Change-Id: I93c4eb78e7cf7c3b589a5c6d04643992667bd69d
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>