Fix window focus after closing preferences dialog of commit dialog
No shells were being passed to the preference dialog when it was
opened from the commit dialog so SWT is forced to choose a shell
to reactivate after the preferences dialog is closed. The chosen
shell becomes the workbench window's shell when the focus should
instead be granted back to the focus dialog. By passing the commit
dialog's shell to the preference dialog, we maintain this hierarchy
and the focus is given back to the commit dialog after the
preferences dialog has been closed.

Change-Id: I4ad646363b13d2134e99d39d9d54a5dae843d8bd
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>