Fix the history page to not override its parent composite's layout
The problem here is that createMainPanel(Composite) was arbitrarily
changing the parent composite's layout. This composite was passed in
by the framework and its layout should not be modified. By changing
its layout to become a StackLayout, the framework was no longer able
to interchange the contents of the view between the different
subclasses of HistoryPage.

The fix is to create a new top level composite that will contain all
the controls that the GitHistoryPage needs in addition to the layout
necessary for the EGit implementation to change what it wants to be
visible to the end user at a given moment in time.

Change-Id: Idd20503faa85aabe613fcece18443a3ead7a86b5