Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Good change.

But the UI tests using the tag dialog do not work anymore on my machine.

The reason is that the OK-button of the create tag dialog never is enabled.

For example, in GitRepositoriesViewTagHandlingTest.testCreateTags the mthod createTag terminates with a TimeoutException

The reason is that createDialog.bot().textWithLabel(UIText.CreateTagDialog_tagMessage).setText(message) does not do anything.

Patch-set: 3