Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

While the command framework is much more modern, I have some concerns here: the contributions can not be configured by the end user and are always visible. I'm not sure if we can just grab the real estate for the toolbar needed to add five buttons (and how many more would we have in the future)? Looking at CVS, they have only one single button configurable for the toolbar, even though their menu has 13 entries.
So I guess we should go back to actions sets.
I played around a bit but wasn't able yet to get the enablement running, so we need to think about this again.
Besides, I found some other issues: Commit and Fetch From are both enabled in the menu and all buttons in the toolbar are enabled if nothing is selected (e.g. after initial start of the workbench). When selecting a Git-controlled project, the enablement in the menu goes on for all items. When closing the project explorer, only Commit and Fetch From are active, both in the menu and in the toolbar...

Patch-set: 1