Do not preselect files with the assume-unchanged flag
When a file has the flag "assume-unchanged" it should per default not
be in the set of files to be committed.

In native git commit -a does not commit such files, commit -- <path>
however does. The pre-selection in the commit dialog contains the list
of files which are committed with commit -a, filtered by the selection
in the Package Explorer. Therefore the file with the flag
"assume-unchanged" should not be pre-selected in the commit dialog.


Bug: 326213
Change-Id: Ia99371cc77b14b9c3e2e551d620499d164405171
Signed-off-by: Stefan Lay <stefan.lay@sap.com>