Rebase Interoperability: do not auto-abort upon conflict
This depends on JGit change

http://egit.eclipse.org/r/#change,1986

(which actually make the two rebase implementations interoperable).
Once interoperability is reached, EGit can stop and have the user
resolve the conflicts; they can then switch to the command line
to continue. Abort is also provided so that users can abort the
rebase in EGit (only if it is an interactive rebase).

Change-Id: Ie694e09b45c43ee82254231d4f619fbe5510d5eb
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>