Initialize test classes properly
GitCloneWizardTests fail when running isolated because the clone
tries to put the repository into <user_home>/<repo_name>; <user_home>
is the "default-default" for the repository location Preference. The
fix initializes that preference to point to the test directory created
during test setup. This initialization already existed in some test
classes, but was pushed down to LocalRepositoryTestCase; the
GitCloneWizard tests were made to inherit from that class to take
advantage of the initialization.

Change-Id: I742a93c67f4f0f56f7eb328bb8712a50d75b3b1e
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>