Fix cloning a repository with detached HEAD
The clone wizard could not handle remote repositories where the HEAD
is no branch. The SourceBranchPage threw an exception in the
LabelProvider. When the user selected "HEAD" on the
CloneDestinationPage the cloned repository was inconsistent.

Furthermore it is now possible to use the CloneOperation with
arbitrary Refs, not only with refs/heads/*.

Bug: 326693
Change-Id: Ie45f2508879b51d8a9190067d37d4277f6092ffb
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>