Small enhancement to the 'Configure Git Repository' dialog
When sharing a project (Team -> Share) it is required to specify a Git
repository location. In order to create a new repository it is
required to first select a project. But it was not possible to
select the project by clicking on the check box in front of it; the
user had to click somewhere else. With this patch the project also
gets selected by clicking on the check box.

Change-Id: I6764f9f7bab0af1d5e754f0ac6c861b79f8eb582
Signed-off-by: Philipp Thun <philipp.thun@sap.com>