Change the graph table to automatically grab all available space
The commit graph table in the 'History' view currently leaves a lot
of extra unused horizontal space. This causes it to clip the last
column despite the fact that it could expand that column's width
to show all of the content of that column. The fix is to alter the
table's layout to force its columns to always take up the all the
space that's available.

Bug: 331183
Change-Id: I8abf7baa5fa86684f855f364aecad941b531e6e7