Rebase Interoperability: do not auto-abort upon conflict
EGit now stops when a conflict is found during rebase and allows the
user to manipulate the content; they can then swith to the command
line to continue the rebase.
The rebase result dialog also offers the options to abort the current
rebase altogether or to open the merge tool to work out the conflicts.

Change-Id: Ie694e09b45c43ee82254231d4f619fbe5510d5eb
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>