Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; IP review completed

I kind of like that behavior better than the current one, but the problem is that there is no scrollbar anymore that allows me to scroll to the right to see the stuff that doesn't fit into the current width of the table (see screenshot). Is there a way to fix this?

Patch-set: 1