Update patch set 1
Patch Set 1:

Where do I see this screenshot? I don't see a link here or in Bugzilla. Though I think I know what you're talking about.

It's really hard to get it to expand beyond the allocated space because it's "preset" to be 100%. I think you need to resize the first two or last two columns a few times to (somehow) force the final column to be capable of going beyond the allocated space (100%) of the table. I'm not aware of any straightforward workarounds off-hand.

Patch-set: 1