Update patch set 1
Patch Set 1:

The context in which I discovered this was when I tried to start the git clone wizard for a given repository (I know, this is not an official API yet - therefore see bug 332316). Anyway, if I have a system that knows about git repositories (e.g. Gerrit Mylyn Integration or Skalli [1]), I'd want an IDE integration in order to directly clone a repository, instead of messing around with URIs manually.

Having a parameter in a method and then ignoring it is always a bug. The use case shouldn't really matter, I guess...

[1] http://wiki.eclipse.org/Skalli/Proposal

Patch-set: 1