Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Better, but not yet ok ;-)

When a file is changed both in the workspace and in the index, the compare editor for the workspace version compares the workspace version with the index version. However, the label of the index version on the right hand side contains the commit id instead of "cached version".

Patch-set: 2