Do not put user to URI when using http / https
Currently the repository selection page (used by clone wizard and
other wizards) automatically puts a user entered in the user field to
the URI. This is a security hole for http and https usage. Furthermore
configuring http based push does not work because proxy servers often
reject URLs with contained user.
The automatic update of the URI field with the user was disabled for
the protocols http and https.

Change-Id: Iad68286d40061aa92e67c72775563dc7b18af87f
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>