Add cheat sheet for clone wizard
This change contains a cheat sheet for the clone wizard to support
users that are new to EGit. It can be found in Eclipse as follows:
Help > Cheat Sheets... > Team/Git > Cloning a Git Repository.
Furthermore the cheat sheet is also opened when clicking on the help
icon (?) from within the wizard.

Change-Id: If6fd730154a2e59daeff89d96cab2006718d8d9f
Signed-off-by: Philipp Thun <philipp.thun@sap.com>