Checkout: display checkout conflicts in a user-friendly way
Also, if some files could not be deleted during checkout, these files
are displayed to the user so that they have a chance to retry the
deletion.

Depends on JGit change

http://egit.eclipse.org/r/#change,2139

which exposes the list of conflicts and fixes some issues when
checking out a commit or remote tracking branch.

Change-Id: Ic8b491b268dffdc3928a4d8175ed2be0de3fafc6
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>